requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
python-mecab-ko==1.3.7

